move Axis, remove workspace_provider base class re #308
