+++ mslice/models/axis.py
@@ -1,0 +1,15 @@
+class Axis(object):
+    def __init__(self, units, start, end, step):
+        self.units = units
+        self.start = start
+        self.end = end
+        self.step = step
+
+    def __eq__(self, other):
+        # This is required for Unit testing
+        return self.units == other.units and self.start == other.start and self.end == other.end \
+            and self.step == other.step and isinstance(other, Axis)
+
+    def __repr__(self):
+        info = (self.units, self.start, self.end, self.step)
+        return "Axis(" + " ,".join(map(repr, info)) + ")"
