+++ mslice/presenters/cut_presenter.py
@@ -1,11 +1,12 @@
 from __future__ import (absolute_import, division, print_function)
 
 from .busy import show_busy
+from mslice.models.axis import Axis
 from mslice.models.cut.cut_algorithm import CutAlgorithm
 from mslice.models.cut.cut_plotter import CutPlotter
-from mslice.models.workspacemanager.mantid_workspace_provider import get_workspace_handle
+from mslice.models.workspacemanager.workspace_provider import get_workspace_handle
 from mslice.presenters.presenter_utility import PresenterUtility
-from mslice.presenters.slice_plotter_presenter import Axis
+
 from mslice.views.cut_view import CutView
 from mslice.widgets.cut.command import Command
 from .validation_decorators import require_main_presenter
